Add branching support.
Added cp and cpi instructions to AVRInstrInfo.td, and also changed the template used for jumping from j to br
Added condition codes for AVR and modified assembly printer to write it out correctly.
Added SREG to register info, as condition codes and branching instructions depend on it.
Added code to analyze, remove and add branches in AVRInstrInfo.