Fix indexed loads and stores off Y. Also fix setting the stack pointer.
Fix offset calculation when using frame pointers. The offset provided by LLVM is calculated off the the value of the stack at function entry, and is a negative value as the stack grows down. To convert it to a indexed load store off Y, we need to invert the sign and add 2 to account for the pushed frame pointer (Y).

Allocating space on the stack by setting the stack pointer was only manipulating 0x3D. It now uses both 0x3D and 0x3E to subtract the required number of bytes and store it back.