Fix code generation with frame pointer omission.
The AVR doesn't have a dedicated SP register that can be used for indexed load and store. Therefore, storing and loading stack slots with SP as the base ptr will lead to terribly bloated code - Every stack reference will involve loading 0x3E:0x3D into an indexed register, followed by and std or ldd.
Instead, the Y register setup code in the prologue is generated if there are stack slots. Reading and writing to stack slots continue to use the Y register, as in functions compiled with a frame pointer. The only difference (in functions that require stack slots) between fpo'd code and non fpo'd code is the pushing and popping of the Y register on entry and exit.